# backend/shadowvale_be/docs/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ShadowVale.API | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/solver-configurations": {
      "get": {
        "tags": [
          "SolverConfigurations"
        ],
        "parameters": [
          {
            "name": "Family",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Algorithm",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "IsActive",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolverConfigurationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolverConfigurationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SolverConfigurationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "SolverConfigurations"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSolverConfigurationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSolverConfigurationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSolverConfigurationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/solver-configurations/{id}": {
      "get": {
        "tags": [
          "SolverConfigurations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "SolverConfigurations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolverConfigurationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolverConfigurationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSolverConfigurationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "SolverConfigurations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/solver-configurations/{id}/clone": {
      "post": {
        "tags": [
          "SolverConfigurations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneSolverConfigurationRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CloneSolverConfigurationRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CloneSolverConfigurationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/solver-configurations/{id}/active": {
      "patch": {
        "tags": [
          "SolverConfigurations"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSolverConfigurationActiveRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSolverConfigurationActiveRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetSolverConfigurationActiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SolverConfigurationDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloneSolverConfigurationRequest": {
        "required": [
          "code",
          "name"
        ],
        "type": "object",
        "properties": {
          "code": {
            "maxLength": 64,
            "pattern": "^[a-z][a-z0-9_]*$",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "type": "string"
          }
        }
      },
...
      "CreateSolverConfigurationRequest": {
        "required": [
          "code",
          "name",
          "algorithm"
        ],
        "type": "object",
        "properties": {
          "code": {
            "maxLength": 64,
            "pattern": "^[a-z][a-z0-9_]*$",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "algorithm": {
            "type": "string"
          },
          "library": {
            "maxLength": 50,
            "type": [
              "null",
              "string"
            ]
          },
          "params": {
            "type": [
              "null",
              "object"
            ]
          },
          "quboWeights": {
            "type": [
              "null",
              "object"
            ]
          },
          "timeBudgetMs": {
            "maximum": 10000,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "JsonElement": { },
...
      "SetSolverConfigurationActiveRequest": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "SolverConfigurationDto": {
        "required": [
          "id",
          "code",
          "name",
          "algorithm",
          "family",
          "variant",
          "library",
          "params",
          "quboWeights",
          "timeBudgetMs",
          "isActive",
          "isAbArm",
          "sessionCount",
          "resultCount",
          "createdAt",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "code": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "algorithm": {
            "type": "string"
          },
          "family": {
            "type": "string"
          },
          "variant": {
            "type": "string"
          },
          "library": {
            "type": [
              "null",
              "string"
            ]
          },
          "params": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "quboWeights": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "timeBudgetMs": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "isAbArm": {
            "type": "boolean"
          },
          "sessionCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "resultCount": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "UpdateSolverConfigurationRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 100,
            "type": "string"
          },
          "library": {
            "maxLength": 50,
            "type": [
              "null",
              "string"
            ]
          },
          "params": {
            "type": [
              "null",
              "object"
            ]
          },
          "quboWeights": {
            "type": [
              "null",
              "object"
            ]
          },
          "timeBudgetMs": {
            "maximum": 10000,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Access token from POST /api/auth/login",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "GameKey": {
        "type": "apiKey",
        "description": "Key built into the game (config Game:ApiKeys)",
        "name": "X-Game-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}